+++ Cargo.toml
@@ -17,6 +17,7 @@ ccitt-t4-t6 = { path = "crates/ccitt" }
 pdf-create = { path = "crates/pdf" }
 sdo-pdf = { path = "crates/sdo-pdf" }
 sdo-ps = { path = "crates/sdo-ps" }
+sdo-util = { path = "crates/sdo-util" }
 signum = { path = "crates/signum", features = ["image"] }
 pdf = "0.7"
 ron = "0.6"
