+++ crates/sdo-util/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "sdo-util"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+signum = { path = "../signum" }
