+++ crates/sdo-web/Cargo.toml
@@ -15,3 +15,4 @@ serde-wasm-bindgen = "0.4.5"
 console_log = { version = "1.0.0", features = ["wasm-bindgen"] }
 log = "0.4.17"
 image = { version = "0.24", default-features = false, features = ["png"] }
+bstr = "1.4.0"
