+++ crates/sdo-web/index.html
@@ -1,63 +1,51 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta content="text/html;charset=utf-8" http-equiv="Content-Type"/>
-    <style>
-      img {
-        border: 1px dotted blue;
-      }
-    </style>
-  </head>
-  <body>
-    <input id="upload" type="file" accept=".sdo, .p24, .e24, .p09">
-    <section id="0001"></section>
-    <section id="cset"></section>
-    <section id="hcim"></section>
 
-    <!-- Note the usage of `type=module` here as this is an ES6 module -->
-    <script type="module">
-      // Use ES module import syntax to import functionality from the module
-      // that we have compiled.
-      //
-      // Note that the `default` import is an initialization function which
-      // will "boot" the module and make it ready to use. Currently browsers
-      // don't support natively imported WebAssembly as an ES module, but
-      // eventually the manual initialization won't be required!
-      import init, { setup } from './pkg/sdo_web.js';
+<head>
+  <meta content="text/html;charset=utf-8" http-equiv="Content-Type" />
+  <style>
+    img {
+      border: 1px dotted blue;
+    }
+  </style>
+</head>
 
-      async function run() {
-        // First up we need to actually load the wasm file, so we use the
-        // default export to inform it where the wasm file is located on the
-        // server, and then we wait on the returned promise to wait for the
-        // wasm to be loaded.
-        //
-        // It may look like this: `await init('./pkg/without_a_bundler_bg.wasm');`,
-        // but there is also a handy default inside `init` function, which uses
-        // `import.meta` to locate the wasm file relatively to js file.
-        //
-        // Note that instead of a string you can also pass in any of the
-        // following things:
-        //
-        // * `WebAssembly.Module`
-        //
-        // * `ArrayBuffer`
-        //
-        // * `Response`
-        //
-        // * `Promise` which returns any of the above, e.g. `fetch("./path/to/wasm")`
-        //
-        // This gives you complete control over how the module is loaded
-        // and compiled.
-        //
-        // Also note that the promise, when resolved, yields the wasm module's
-        // exports which is the same as importing the `*_bg` module in other
-        // modes
-        await init();
+<body>
+  <input id="upload" type="file" accept=".sdo, .p24, .e24, .p09">
+  <section id="0001"></section>
+  <section id="cset"></section>
+  <section id="hcim"></section>
 
-        window.__x = setup("upload");
+  <script type="module">
+    import init, { Handle } from './pkg/sdo_web.js';
+
+    async function run() {
+      await init();
+
+      const h = new Handle(document);
+
+      async function runFile(file) {
+        const buf = await file.arrayBuffer();
+        const arr = new Uint8Array(buf);
+        h.do_stuff(file.name, arr);
       }
 
-      run();
-    </script>
-  </body>
+      async function runFiles(field) {
+        for (const file of field.files) {
+          await runFile(file);
+          console.log(`Completed '${file.name}'`);
+        }
+      }
+      
+      const inputField = document.getElementById("upload");
+      inputField.addEventListener('change', (event) => {
+        runFiles(event.target);
+      });
+      runFiles(inputField);
+    }
+
+    run();
+  </script>
+</body>
+
 </html>
